Various Bug Fixes
Fixed an issue where the kid in Pewter City blocking the path to Route 3
would disappear aftering beating Brock regardless of the condition.

Removed the ledge above Diglett's Cave on Route 2 and restored the ledge
on Route 4 that blocks the way back into the route.

Fixed an issue where beating Giovanni in Viridian Gym would remove the
Rocket Grunts in front of his room in Rocket Hideout, making it
impossible to open the door there.

Swapped the Seafoam Island entrances.

Reduced the sight of a trainer on Route 25 to prevent being able to get
the item behind the cut tree without cut.

Added GO HOME option to the main menu that teleports you back to Pallet
Town.

Fixed an issue where the game would softlock in teh next battle after
the player finished a double battle.

Changed it so that when choosing if you want to send out a new Pokemon
you can still see what Pokemon the opponent intends to send out.

Added dummy moves for every Pokemon so that Pokemon can start with four
moves.