Implemented Trainersanity
Implemented a system where items can be given to the player after
beating a trainer.

Made it so that trainers won't disappear if the Trainersanity AP option
is on.

Added Route 4 and Route 10 to the free fly options.

Split the Rival fight on Route 22 back into two fights.

Added symbols for static Pokemons levels.

Made it so that you can't fish until you have your first Pokemon.

Updated the extractor tool to create a single extracted data file with
all 4 versions in it.