Added Alternate Layouts for Route 10 & 23
Added an option to add a waterfall to Route 10.

Added an option to add a waterfall to Route 23.

Made it so that you can't go down waterfalls without being able to use
the Waterfall field move.

Made it so you actually need to have a Pokemon that knows flash to
traverse dark caves with required flash instead of just having the HM.