Make Berry Pouch and TM Case Starting Items
Add the Berry Pouch and TM Case to the player's bag at the start of the game.

Make it so that getting a berry doesn't add the Berry Pouch to the player's bag.

Make it so that getting a HM/TM doesn't add the TM Case to the player's bag.